Resolve DEP-102 review findings
