+++ .env.sample
@@ -62,6 +62,19 @@ MSPRESIDIO_ANONYMIZER_URL=
 # tesseract binary. The sidecar decides CPU vs GPU on its side.
 RDP_OCR_SERVER_URL=
 
+# Embedded OCR image settings. The GPU image already supplies
+# RDP_OCR_TRANSPORT=resident-rdp and OCR_GPU_PREPROCESS=1; the CPU
+# compatibility image supplies RDP_OCR_TRANSPORT=bmp. Override only when
+# diagnosing image/transport parity.
+# RDP_OCR_TRANSPORT=resident-rdp
+# OCR_GPU_PREPROCESS=1
+# Resident GPU capacity and cache bounds:
+# OCR_RESIDENT_MAX_SESSIONS=32
+# OCR_RESIDENT_VRAM_MB=2048
+# OCR_RESIDENT_TTL_SECONDS=300
+# OCR_RESIDENT_CACHE_ENTRIES=128
+# OCR_RESIDENT_RECOGNITION_CACHE_ENTRIES=512
+
 # action the realtime agent-side PII guard takes on detection:
 # redact (default: blank the PII and continue) | kill (drop and terminate) | redact_and_kill
 RDP_PII_GUARD_POLICY=
