+++ gateway/models/rdp_entity_detection.go
@@ -1,11 +1,5 @@
 package models
 
-import (
-	"context"
-
-	"gorm.io/gorm"
-)
-
 // RDPEntityDetection represents a single PII entity detected in an RDP session frame.
 // Coordinates (X, Y, Width, Height) are in screen-space pixels, matching the canvas
 // coordinate system used by the RDP session replay player.
@@ -31,70 +25,6 @@ func BulkInsertRDPEntityDetections(detections []RDPEntityDetection) error {
 	return DB.Table("private.rdp_entity_detections").
 		Omit("id").
 		CreateInBatches(detections, 100).Error
-}
-
-// PersistRDPGuardrailViolation appends one report and its detections in a
-// transaction. The session row is locked and report_id is checked inside its
-// guardrails_info JSON array, making an agent retry idempotent without a
-// separate side table.
-func PersistRDPGuardrailViolation(
-	ctx context.Context,
-	orgID, sessionID, reportID string,
-	info []byte,
-	detections []RDPEntityDetection,
-) error {
-	return DB.WithContext(ctx).Transaction(func(tx *gorm.DB) error {
-		var lockedSessionID string
-		lock := tx.Raw(
-			`SELECT id FROM private.sessions WHERE org_id = ? AND id = ? FOR UPDATE`,
-			orgID,
-			sessionID,
-		).Scan(&lockedSessionID)
-		if lock.Error != nil {
-			return lock.Error
-		}
-		if lock.RowsAffected == 0 {
-			return ErrNotFound
-		}
-
-		var duplicate bool
-		if err := tx.Raw(
-			`SELECT EXISTS (
-				SELECT 1
-				FROM jsonb_array_elements(COALESCE(guardrails_info, '[]'::jsonb)) AS item
-				WHERE item->>'report_id' = ?
-			)
-			FROM private.sessions
-			WHERE org_id = ? AND id = ?`,
-			reportID,
-			orgID,
-			sessionID,
-		).Scan(&duplicate).Error; err != nil {
-			return err
-		}
-		if duplicate {
-			return nil
-		}
-
-		update := tx.Table("private.sessions").
-			Where("org_id = ? AND id = ?", orgID, sessionID).
-			Update(
-				"guardrails_info",
-				gorm.Expr("COALESCE(guardrails_info, '[]'::jsonb) || ?::jsonb", info),
-			)
-		if update.Error != nil {
-			return update.Error
-		}
-		if update.RowsAffected == 0 {
-			return ErrNotFound
-		}
-		if len(detections) == 0 {
-			return nil
-		}
-		return tx.Table("private.rdp_entity_detections").
-			Omit("id").
-			CreateInBatches(detections, 100).Error
-	})
 }
 
 // GetRDPEntityDetections returns all entity detections for a session, ordered by frame index.
